Task #1884 Add time traces to deduper
Created a new module called timer. This module will deal with times in deduper running.
Created too a struct called TIMES, which will store some times. Now the interface is
creating the struct and the model runner puts the values in it. The other changes regard
to adding the new module to the project.